+++ Pages/Management/ProductManagement.cshtml.cs
@@ -1,3 +1,4 @@
+using MDP.DevKit.LineMessaging;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Web0524.Models;
@@ -8,11 +9,13 @@ namespace Web0524.Pages.Management
     {
         private readonly IProductService _productService;
         private readonly IPgroupService _pgroupService;
+        private readonly IUserService _userService;
 
-        public ProductManagementModel(IProductService productService, IPgroupService pgroupService)
+        public ProductManagementModel(IProductService productService, IPgroupService pgroupService, IUserService userService)
         {
             _productService = productService;
             _pgroupService = pgroupService;
+            _userService = userService;
         }
 
         [BindProperty]
@@ -25,8 +28,12 @@ namespace Web0524.Pages.Management
 
         public List<Pgroup> PGroups { get; set; } = new(); // 為顯示 PGname
         [IgnoreAntiforgeryToken]
-        public void OnGet(int? id)
+        public IActionResult OnGet(int? id)
         {
+
+            var check = _userService.CheckCurrentUserPermission(this);
+            if (check != null) return check;
+
             Products = _productService.GetAllProducts().ToList();
             PGroups = _pgroupService.GetAllPgroups().ToList(); ;
 
@@ -38,13 +45,31 @@ namespace Web0524.Pages.Management
                     Product = product;
                 }
             }
+            return Page();
         }
         [IgnoreAntiforgeryToken]
         public JsonResult OnPostSave()
         {
+
+            var pageName = this.GetType().Name.Replace("Model", "");
+            var userId = User.FindFirst(System.Security.Claims.ClaimTypes.Sid)?.Value;
+            if (string.IsNullOrEmpty(userId) || !_userService.HasPagePermissionByName(userId, pageName))
+            {
+                return new JsonResult(new { success = false, message = "無權限" });
+            }
             if (!ModelState.IsValid)
             {
                 return new JsonResult(new { success = false, message = "請填寫所有必要欄位。" });
+            }
+
+            if (Product.ProductId > 0)
+            {
+                // 取得原本的圖片資料
+                var original = _productService.GetProductById(Product.ProductId);
+                if (original != null && PhotoFile == null)
+                {
+                    Product.Photo = original.Photo; // 沒有上傳新圖，就保留原圖
+                }
             }
 
             if (PhotoFile != null)
@@ -65,9 +90,16 @@ namespace Web0524.Pages.Management
                 return new JsonResult(new { success = true, message = "產品已成功更新。" });
             }
         }
+
         [IgnoreAntiforgeryToken]
         public JsonResult OnPostDelete(int id)
         {
+            var pageName = this.GetType().Name.Replace("Model", "");
+            var userId = User.FindFirst(System.Security.Claims.ClaimTypes.Sid)?.Value;
+            if (string.IsNullOrEmpty(userId) || !_userService.HasPagePermissionByName(userId, pageName))
+            {
+                return new JsonResult(new { success = false, message = "無權限" });
+            }
             _productService.DeleteProduct(id);
             return new JsonResult(new { success = true, message = "產品已刪除。" });
         }
@@ -75,6 +107,12 @@ namespace Web0524.Pages.Management
         [IgnoreAntiforgeryToken]
         public IActionResult OnGetImage(int id)
         {
+            var pageName = this.GetType().Name.Replace("Model", "");
+            var userId = User.FindFirst(System.Security.Claims.ClaimTypes.Sid)?.Value;
+            if (string.IsNullOrEmpty(userId) || !_userService.HasPagePermissionByName(userId, pageName))
+            {
+                return new JsonResult(new { success = false, message = "無權限" });
+            }
             Products = _productService.GetAllProducts().ToList();
             var product = Products.FirstOrDefault(p => p.ProductId == id);
             if (product?.Photo != null)
